+++ R/subsetted.R
@@ -1,35 +1,3 @@
-countries1 <- countries
-names(countries1)= tolower(countries1)
-
-#' Get country states
-#'
-#' Fetch all country states
-#'
-#' @rdname countryitems
-#' @format NULL
-#' @docType NULL
-#' @keywords NULL
-#' @export
-getCTRY<-lapply(countries1,function(i){
-  rlang::new_function(
-    args = rlang::exprs(item = c("currency","language","callingcode","population"), as.of = 2023),
-    rlang::expr({
-      ctry.name = !!i
-      item = match.arg(item)
-      print(item)
-      print("heelo")
-    }),
-    env = asNamespace("quickcode")
-  )
-})
-
-
-# xxy<-lapply(countries1,function(i){
-#   p = i
-#   paste0("popeplr",p)
-# })
-
-
 #' Fetch countries that start with specified characters
 #'
 #' Subset and return countries given a specified characters to search
@@ -308,47 +276,4 @@ byContinent <-function(name = c('asia','europe','africa','north america','south 
     return(NULL)
   }
 }
-
-#' @export
-languageOf <- lapply(countries1, function(cnt){
-  unlist(data.pck.cntry[data.pck.cntry$name==cnt,]$officiallanguage)
-})
-
-#' @export
-capitalOf <- lapply(countries1, function(cnt){
-  unlist(data.pck.cntry[data.pck.cntry$name==cnt,]$capital)
-})
-
-#' @export
-populationOf <- lapply(countries1, function(cnt){
-  unlist(data.pck.cntry[data.pck.cntry$name==cnt,]$population2023)
-})
-
-#' @export
-callingCodeOf <- lapply(countries1, function(cnt){
-  paste0("+",unlist(data.pck.cntry[data.pck.cntry$name==cnt,]$callingcode))
-})
-
-#' @export
-currencyOf <- lapply(countries1, function(cnt){
-  dcntry <- data.pck.cntry[data.pck.cntry$name==cnt,c("currency","symbol","isocode","fractionalunity")]
-  if(nrow(dcntry)) unlist(dcntry) else NULL
-})
-
-
-city_time1 <- unlist(unique(city_time$City))
-names(city_time1) <- city_time1
-
-#' @export
-timeOf <- lapply(city_time1, function(cnt){
-  #message("!!! does not take daylight savings into account yet")
-  us.time <- as.POSIXct(Sys.time(), tz = "America/New_York")
-  as.POSIXct(as.numeric(unlist(city_time[city_time$City == cnt,]$Timediff)[1]) + us.time)
-})
-
-
-#' @export
-continentOf <- lapply(countries1, function(cnt){
-  unlist(data.pck.cntry[data.pck.cntry$name==cnt,]$continent)
-})