+++ inst/.temp/compiler.R
@@ -174,3 +174,26 @@ use_data(city_time,overwrite = T)
 use_data(country_names,overwrite = T)
 use_data(country_money,overwrite = T)
 use_data(country_language,overwrite = T)
+
+
+#' #' Get country states
+#' #'
+#' #' Fetch all country states
+#' #'
+#' #' @rdname countryitems
+#' #' @format NULL
+#' #' @docType NULL
+#' #' @keywords NULL
+#' #' @export
+#' getCTRY<-lapply(countries1,function(i){
+#'   rlang::new_function(
+#'     args = rlang::exprs(item = c("currency","language","callingcode","population"), as.of = 2023),
+#'     rlang::expr({
+#'       ctry.name = !!i
+#'       item = match.arg(item)
+#'       print(item)
+#'       print("heelo")
+#'     }),
+#'     env = asNamespace("quickcode")
+#'   )
+#' })