+++ tests/testthat/test-countries1.R
@@ -1,0 +1,3 @@
+test_that("Is country list intact", {
+  expect_equal(length(countries1) > 100, TRUE)
+})