+++ tests/testthat/test-data.R
@@ -1,0 +1,3 @@
+test_that("Test combined dataset", {
+  expect_equal(nrow(data.pck.cntry) > 100, TRUE)
+})