+++ NEWS.md
@@ -1,11 +1,9 @@
 # r2country 1.0
 
-## Features
-
  - [x] All country names
  - [x] All country currencies
  - [x] All country area code
  - [x] All country current time
- - [x] All country presidents
+ - [x] All country capitals
  - [x] All country population
  - [x] All country languages